general: Add service/multi and response variant
The response variant allows the result of a protocol message to be one
of several types of response: Message, Vec<Message> or bytes. A vec of
messages is considered a stream of chunks while bytes is considered a
raw response.

When a service/multi is necessary for transmission, the protocol handler
converts the header, chunks and tail into bytes and passes that back
out. The idea is that in the future it can just be streamed.